move moveObject reducer to separate module
