+++ src/app/features/editor/slice.ts
@@ -1,15 +1,16 @@
 import type { PayloadAction } from "@reduxjs/toolkit";
-import { createSlice, original } from "@reduxjs/toolkit";
+import { createSlice } from "@reduxjs/toolkit";
 import produce from "immer";
 import type {
-  Event,
   Card,
   Deck,
-  Property,
-  Middleware,
-  GameDefinition,
+  Event,
   Game,
+  GameDefinition,
+  Middleware,
+  Property,
 } from "../../../api/services/game/types";
+import { propertyValue } from "../../../api/services/game/types";
 import {
   createEntityReducerFactory,
   createId,
@@ -20,7 +21,6 @@ import {
   addNodeBySplitting,
   removeNodeByKey,
 } from "../../../lib/reactMosaicExtensions";
-import { propertyValue } from "../../../api/services/game/types";
 import type {
   EditorObjectId,
   EditorState,
@@ -32,7 +32,7 @@ import type {
 import { editorObjectIdType, panelLayoutType } from "./types";
 import { defaultPanelLayout } from "./panels/defaultPanelLayout";
 import { selectors } from "./selectors";
-import { createObjectByIdPredicate, selectedList } from "./utils/objectUtils";
+import { moveObject } from "./reducers/moveObject";
 
 const panelStorage = createZodStorage(
   panelLayoutType,
@@ -208,53 +208,7 @@ const editorSlice = createSlice({
         ...payload,
       });
     },
-    moveObject(
-      state,
-      {
-        payload: [movedObjectId, targetObjectId],
-      }: PayloadAction<[EditorObjectId, EditorObjectId]>
-    ) {
-      if (!state.game) {
-        return;
-      }
-
-      if (movedObjectId.type === "card" && targetObjectId.type === "deck") {
-        const { cards, decks } = state.game.definition;
-        const card = cards.find((c) => c.cardId === movedObjectId.cardId);
-        const deck = decks.find((d) => d.deckId === targetObjectId.deckId);
-        if (!card || !deck) {
-          throw new Error("Unknown card or deck");
-        }
-        card.deckId = deck.deckId;
-        return;
-      }
-
-      if (movedObjectId.type !== targetObjectId.type) {
-        throw new Error(
-          "Cannot move an object to a target of a different type"
-        );
-      }
-
-      const list = selectedList(state, movedObjectId.type);
-      if (!list) {
-        throw new Error(
-          `Object type "${movedObjectId.type}" does not support moving`
-        );
-      }
-
-      const idx1 =
-        original(list)?.findIndex(createObjectByIdPredicate(movedObjectId)) ??
-        -1;
-      const idx2 =
-        original(list)?.findIndex(createObjectByIdPredicate(targetObjectId)) ??
-        -1;
-      if (idx1 === -1 || idx2 === -1) {
-        throw new Error("Could not find target or object to move");
-      }
-
-      const [moved] = list.splice(idx1, 1);
-      list.splice(idx2, 0, moved);
-    },
+    moveObject,
     createProperty(
       state,
       {
